vim linking
I'd like to have commands then vim --servername <name> --remote and
want to go through my files and not doing that again with vim. It is of
cause not right to do so all the time. To do so I redefine cd and vim
only when I start a vimlink to a specific server and bring it back to
default when the link shall perish. $VIMLINKED is the environment
variable that helps me to do so. It holds the current bound servername
or an emtpy string.